+++ RMSD/rmsd.h
@@ -4,10 +4,10 @@
 
 #include "point.h"
 
-double compute_rmsd_cpu(std::vector<point*> s1, std::vector<point*> s2);
+float compute_rmsd_cpu(std::vector<point*> s1, std::vector<point*> s2);
 
-int sumPointDistances(std::vector<point*> s1, std::vector<point*> s2);
+float sumPointDistances(std::vector<point*> s1, std::vector<point*> s2);
 
-int distanceBetweenTwoPoints(point* p1, point* p2);
+float distanceBetweenTwoPoints(point* p1, point* p2);
 
-int distanceBetweenTwoCoordsSquared(int c1, int c2);
+float distanceBetweenTwoCoordsSquared(float c1, float c2);
